refactor(protocol): move OPEN negotiation (ValidateOpen/OpenNegotiation) into BGPLite.Protocol (#269)
OPEN validation + capability negotiation is pure RFC 4271 §4.2/§6.2/§6.2.2
logic (version, 4-octet-ASN capability, expected ASN, hold time
min(local,peer), BGP Identifier) with no I/O — yet it lived as internal
static on BgpSession, unreachable without session machinery. As part of
extracting BGPLite.Protocol into a standalone library, the rules now
live in the protocol assembly:

- new public OpenNegotiator.Validate + OpenNegotiation record (moved
  verbatim; XML docs updated to point at the new home)
- BgpSession.ValidateOpen reduces to: call the negotiator, apply the
  negotiated values, fire peer-identification + GR logging
- BgpSessionValidateOpenTests renamed OpenNegotiatorTests, calling the
  public API (no InternalsVisibleTo hop); all branches unchanged

No behavior change — the same BgpNotificationException codes/subcodes
are thrown on the same inputs.
